refactor: repository 패키지를 application/port로 통합 및 ClickHouse 쿼리 최적화
Repository Port 인터페이스를 기존 repository 패키지에서 application/port 패키지로 이동하여
API Port와 Repository Port를 하나의 port 패키지로 통합. ClickHouse 데이터 적재 쿼리에서
NOT IN (SELECT DISTINCT ...) 패턴을 NOT EXISTS로 변경하고, match_final_item 적재 시
임시 테이블을 활용한 단계별 로컬 적재 방식으로 개선.
