+++ infra/persistence/src/main/java/com/mmrtr/lol/infra/persistence/league/repository/TierCutoffRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.mmrtr.lol.infra.persistence.league.repository;
 
 import com.mmrtr.lol.domain.league.domain.TierCutoff;
-import com.mmrtr.lol.domain.league.repository.TierCutoffRepositoryPort;
+import com.mmrtr.lol.domain.league.application.port.TierCutoffRepositoryPort;
 import com.mmrtr.lol.infra.persistence.league.entity.TierCutoffEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
