+++ infra/persistence/src/main/java/com/mmrtr/lol/infra/persistence/league/scheduler/SummonerRankingCalculationService.java
@@ -3,8 +3,8 @@ package com.mmrtr.lol.infra.persistence.league.scheduler;
 import com.mmrtr.lol.common.type.Platform;
 import com.mmrtr.lol.domain.league.domain.SummonerRanking;
 import com.mmrtr.lol.domain.league.domain.TierCutoff;
-import com.mmrtr.lol.domain.league.repository.SummonerRankingRepositoryPort;
-import com.mmrtr.lol.domain.league.repository.TierCutoffRepositoryPort;
+import com.mmrtr.lol.domain.league.application.port.SummonerRankingRepositoryPort;
+import com.mmrtr.lol.domain.league.application.port.TierCutoffRepositoryPort;
 import com.mmrtr.lol.domain.league.application.usecase.SaveTierCutoffUseCase;
 import com.mmrtr.lol.infra.persistence.league.repository.LeagueSummonerJpaRepository;
 import com.mmrtr.lol.infra.persistence.league.repository.MostChampionProjection;
