+++ infra/rabbitmq/src/main/java/com/mmrtr/lol/infra/rabbitmq/listener/MatchFindListener.java
@@ -1,6 +1,6 @@
 package com.mmrtr.lol.infra.rabbitmq.listener;
 
-import com.mmrtr.lol.domain.match.repository.MatchRepositoryPort;
+import com.mmrtr.lol.domain.match.application.port.MatchRepositoryPort;
 import com.mmrtr.lol.domain.match.application.port.MatchApiPort;
 import com.mmrtr.lol.infra.rabbitmq.config.RabbitMqBinding;
 import com.mmrtr.lol.infra.rabbitmq.dto.SummonerRenewalMessage;
